add utility module (contains ModelSyncer)
